+++ shapiq/explainer/_base.py
@@ -17,7 +17,9 @@ class Explainer:
     Args:
         model: The model object to be explained.
         data: A background dataset to be used for imputation in ``TabularExplainer``.
-        class_label: The class label to be explained. Defaults to ``None``.
+        class_index: The class index of the model to explain. Defaults to ``None``, which will set
+            the class index to ``1`` per default for classification models and is ignored for
+            regression models.
         **kwargs: Additional keyword-only arguments passed to ``TabularExplainer`` or ``TreeExplainer``.
 
     Attributes:
@@ -26,12 +28,12 @@ class Explainer:
     """
 
     def __init__(
-        self, model, data: Optional[np.ndarray] = None, class_label: Optional[int] = None, **kwargs
+        self, model, data: Optional[np.ndarray] = None, class_index: Optional[int] = None, **kwargs
     ) -> None:
 
         self._model_class = print_class(model)
         self._predict_function, self._model_type = get_predict_function_and_model_type(
-            model, self._model_class, class_label
+            model, self._model_class, class_index
         )
         self.model = model
 
@@ -55,7 +57,7 @@ class Explainer:
             if self._model_type in list(get_explainers()):
                 _explainer = get_explainers()[self._model_type]
                 self.__class__ = _explainer
-                _explainer.__init__(self, model=model, data=data, class_label=class_label, **kwargs)
+                _explainer.__init__(self, model=model, data=data, class_index=class_index, **kwargs)
 
     def explain(self, x: np.ndarray) -> InteractionValues:
         """Explain the model's prediction in terms of interaction values.