+++ shapiq/explainer/tree/explainer.py
@@ -34,7 +34,9 @@ class TreeExplainer(Explainer):
         index: The type of interaction to be computed. It can be one of
             ``["k-SII", "SII", "STII", "FSII", "BII", "SV"]``. All indices apart from ``"BII"`` will
             reduce to the ``"SV"`` (Shapley value) for order 1. Defaults to ``"k-SII"``.
-        class_label: The class label of the model to explain.
+        class_index: The class index of the model to explain. Defaults to ``None``, which will set
+            the class index to ``1`` per default for classification models and is ignored for
+            regression models.
     """
 
     def __init__(
@@ -43,7 +45,7 @@ class TreeExplainer(Explainer):
         max_order: int = 2,
         min_order: int = 1,
         index: str = "k-SII",
-        class_label: Optional[int] = None,
+        class_index: Optional[int] = None,
         **kwargs,
     ) -> None:
 
@@ -57,7 +59,7 @@ class TreeExplainer(Explainer):
             index = "SV"
 
         # validate and parse model
-        validated_model = validate_tree_model(model, class_label=class_label)
+        validated_model = validate_tree_model(model, class_label=class_index)
         self._trees: list[TreeModel] = copy.deepcopy(validated_model)
         if not isinstance(self._trees, list):
             self._trees = [self._trees]
@@ -65,7 +67,7 @@ class TreeExplainer(Explainer):
 
         self._max_order: int = max_order
         self._min_order: int = min_order
-        self._class_label: Optional[int] = class_label
+        self._class_label: Optional[int] = class_index
 
         # setup explainers for all trees
         self._treeshapiq_explainers: list[TreeSHAPIQ] = [