+++ shapiq/games/imputer/base.py
@@ -47,7 +47,7 @@ class Imputer(Game):
         random_state: Optional[int] = None,
     ) -> None:
         if callable(model) and not hasattr(model, "_predict_function"):
-            self._predict_function = utils.ModelPredictor.predict_callable
+            self._predict_function = utils.predict_callable
         else:  # shapiq.Explainer adds a predict function to the model to make it callable
             self._predict_function = model._predict_function
         self.model = model