+++ tests/tests_explainer/test_explainer_tabular_models.py
@@ -18,7 +18,7 @@ def test_torch_reg(torch_reg_model, background_reg_data):
     values = explainer.explain(x_explain)
     assert isinstance(values, InteractionValues)
     sum_of_values = sum(values.values) + values.baseline_value
-    assert prediction == pytest.approx(sum_of_values)
+    assert prediction == pytest.approx(sum_of_values, rel=0.01)
 
 
 def test_torch_clf(torch_clf_model, background_clf_data):
@@ -29,17 +29,17 @@ def test_torch_clf(torch_clf_model, background_clf_data):
     x_explain_tensor = torch.tensor(x_explain, dtype=torch.float32).reshape(1, -1)
     prediction = torch_clf_model(x_explain_tensor).detach().numpy()[0]
 
-    explainer = Explainer(model=torch_clf_model, data=background_clf_data, class_label=2)
+    explainer = Explainer(model=torch_clf_model, data=background_clf_data, class_index=2)
     values = explainer.explain(x_explain)
     assert isinstance(values, InteractionValues)
     sum_of_values = sum(values.values) + values.baseline_value
-    assert prediction[2] == pytest.approx(sum_of_values)
+    assert prediction[2] == pytest.approx(sum_of_values, rel=0.001)
 
-    explainer = Explainer(model=torch_clf_model, data=background_clf_data, class_label=0)
+    explainer = Explainer(model=torch_clf_model, data=background_clf_data, class_index=0)
     values = explainer.explain(x_explain)
     assert isinstance(values, InteractionValues)
     sum_of_values = sum(values.values) + values.baseline_value
-    assert prediction[0] == pytest.approx(sum_of_values)
+    assert prediction[0] == pytest.approx(sum_of_values, rel=0.001)
 
 
 def test_sklearn_clf(rf_clf_model, background_clf_data):
@@ -48,24 +48,24 @@ def test_sklearn_clf(rf_clf_model, background_clf_data):
     x_explain = background_clf_data[0]
     prediction = rf_clf_model.predict_proba(x_explain.reshape(1, -1))[0]
 
-    explainer = TabularExplainer(model=rf_clf_model, data=background_clf_data, class_label=2)
+    explainer = TabularExplainer(model=rf_clf_model, data=background_clf_data, class_index=2)
     values = explainer.explain(x_explain)
     assert isinstance(values, InteractionValues)
     sum_of_values = sum(values.values) + values.baseline_value
-    assert prediction[2] == pytest.approx(sum_of_values)
+    assert prediction[2] == pytest.approx(sum_of_values, rel=0.001)
 
-    explainer = TabularExplainer(model=rf_clf_model, data=background_clf_data, class_label=0)
+    explainer = TabularExplainer(model=rf_clf_model, data=background_clf_data, class_index=0)
     values = explainer.explain(x_explain)
     assert isinstance(values, InteractionValues)
     sum_of_values = sum(values.values) + values.baseline_value
-    assert prediction[0] == pytest.approx(sum_of_values)
+    assert prediction[0] == pytest.approx(sum_of_values, rel=0.001)
 
     # do the same with the bare explainer (only for class_label=2)
-    explainer = Explainer(model=rf_clf_model, data=background_clf_data, class_label=2)
+    explainer = Explainer(model=rf_clf_model, data=background_clf_data, class_index=2)
     values = explainer.explain(x_explain)
     assert isinstance(values, InteractionValues)
     sum_of_values = sum(values.values)
-    assert prediction[2] == pytest.approx(sum_of_values)
+    assert prediction[2] == pytest.approx(sum_of_values, rel=0.001)
 
 
 def test_sklearn_reg(rf_reg_model, background_reg_data):
@@ -94,20 +94,20 @@ def test_sklearn_clf_logistic_regression(lr_clf_model, background_clf_data):
     x_explain = background_clf_data[0]
     prediction = lr_clf_model.predict_proba(x_explain.reshape(1, -1))[0]
 
-    explainer = TabularExplainer(model=lr_clf_model, data=background_clf_data, class_label=2)
+    explainer = TabularExplainer(model=lr_clf_model, data=background_clf_data, class_index=2)
     values = explainer.explain(x_explain)
     assert isinstance(values, InteractionValues)
     sum_of_values = sum(values.values) + values.baseline_value
     assert prediction[2] == pytest.approx(sum_of_values)
 
-    explainer = TabularExplainer(model=lr_clf_model, data=background_clf_data, class_label=0)
+    explainer = TabularExplainer(model=lr_clf_model, data=background_clf_data, class_index=0)
     values = explainer.explain(x_explain)
     assert isinstance(values, InteractionValues)
     sum_of_values = sum(values.values) + values.baseline_value
     assert prediction[0] == pytest.approx(sum_of_values)
 
     # do the same with the bare explainer (only for class_label=2)
-    explainer = Explainer(model=lr_clf_model, data=background_clf_data, class_label=2)
+    explainer = Explainer(model=lr_clf_model, data=background_clf_data, class_index=2)
     values = explainer.explain(x_explain)
     assert isinstance(values, InteractionValues)
     sum_of_values = sum(values.values) + values.baseline_value